+++ benchmarks/test_v3_getbulk.py
@@ -14,9 +14,9 @@ from easysnmp import snmp_bulkwalk as e_snmp_bulkwalk
 
 # Gufo SNMP modules
 from gufo.snmp import SnmpVersion
-from gufo.snmp.async_client import SnmpSession as AsyncSnmpSession
+from gufo.snmp.aio import SnmpSession as AsyncSnmpSession
 from gufo.snmp.snmpd import Snmpd
-from gufo.snmp.sync_client import SnmpSession as SyncSnmpSession
+from gufo.snmp.sync import SnmpSession as SyncSnmpSession
 from gufo.snmp.user import Aes128Key, KeyType, Sha1Key, User
 
 BASE_OID = "1.3.6"