+++ examples/aio/get.py
@@ -1,13 +1,13 @@
 import asyncio
 import sys
 
-from gufo.snmp import SnmpSession
+from gufo.snmp.aio import SnmpSession
 
 
 async def main(addr: str, community: str, oid: str) -> None:
     async with SnmpSession(addr=addr, community=community) as session:
-        async for k, v in session.getnext(oid):
-            print(f"{k}: {v}")
+        r = await session.get(oid)
+        print(r)
 
 
 asyncio.run(main(sys.argv[1], sys.argv[2], sys.argv[3]))